refactor: complete repository/service layer separation
✅ Architecture Changes:
- Split ServiceProvidersRepository into ServiceProvidersRepository + ServiceAccountsRepository
- Eliminated AdminReadRepository and AdminWriteRepository
- Created AdminReadService and AdminWriteService that reuse regular repositories
- Moved admin authorization from repository to service layer

✅ Repository Layer:
- Repositories now contain only database operations (pure data access)
- Added admin-specific methods to existing repositories (get_all_groups, get_all_users, etc.)
- Created AuditLogsRepository for reading audit logs separately from logging operations

✅ Service Layer:
- AdminReadService and AdminWriteService handle authorization in _validate_admin_permissions()
- Services inject regular repositories for database operations
- Clear separation: Services = Business Logic + Auth, Repositories = Data Access

✅ Infrastructure:
- Updated Makefile with background database commands (db_start, db_stop, db_status, db_logs)
- Fixed all import statements for new admin service locations
- All tests passing (24/24) ✓

This achieves proper separation of concerns with Single Responsibility Principle compliance.
