+++ src/routers/admin/pages/service_providers_view.py
@@ -5,7 +5,7 @@ from pydantic import HttpUrl, ValidationError
 from templates.template_manager import Breadcrumb, template_manager
 
 from ....dependencies import get_admin_read_service, get_admin_write_service
-from ....services.admin.write_service import AdminWriteService
+from ....services.admin_write_service import AdminWriteService
 
 router = APIRouter(
     prefix="/service-providers",
