+++ src/tests/integration/test_5_groups_scopes.py
@@ -9,12 +9,12 @@ def test_update_scopes(client):
     new_scopes = "read,write,delete"
     new_contract = "datapass_48"
     response = client.patch(
-        f"/groups/{new_group_data["id"]}/scopes?scopes={new_scopes}&contract_description={new_contract}"
+        f"/groups/{new_group_data['id']}/scopes?scopes={new_scopes}&contract_description={new_contract}"
     )
     assert response.status_code == 200
 
     # Verify access was updated
-    response = client.get(f"/groups/{new_group_data["id"]}")
+    response = client.get(f"/groups/{new_group_data['id']}")
     assert response.status_code == 200
     group = response.json()
     assert group["scopes"] == new_scopes
@@ -22,12 +22,12 @@ def test_update_scopes(client):
 
     new_contract_url = "https://example.com/contract"
     response = client.patch(
-        f"/groups/{new_group_data["id"]}/scopes?scopes={new_scopes}&contract_url={new_contract_url}"
+        f"/groups/{new_group_data['id']}/scopes?scopes={new_scopes}&contract_url={new_contract_url}"
     )
     assert response.status_code == 200
 
     # Verify access was updated
-    response = client.get(f"/groups/{new_group_data["id"]}")
+    response = client.get(f"/groups/{new_group_data['id']}")
     assert response.status_code == 200
     group = response.json()
     assert group["contract_url"] == new_contract_url
